Adding Project entities.

+++ src/Core/Umio.API.Entities/Entidades/Produtos/Acompanhamento.cs
@@ -2,7 +2,7 @@
 {
     public class Acompanhamento : Produto
     {
-        public Acompanhamento(Guid id, string nome, double preco, string descricao, string comentarios, string imagem) : base(id, nome, preco, descricao, comentarios, imagem)
+        public Acompanhamento(Guid id, string nome, decimal preco, string descricao, string comentarios, string imagem) : base(id, nome, preco, descricao, comentarios, imagem)
         {
         }
     }
