+++ src/Core/Umio.API.Entities/Entidades/Produtos/Prato.cs
@@ -2,7 +2,7 @@
 {
     public class Prato : Produto
     {
-        private Prato(Guid id, string nome, double preco, string descricao, string comentarios, string imagem) : base(id, nome, preco, descricao, comentarios, imagem)
+        private Prato(Guid id, string nome, decimal preco, string descricao, string comentarios, string imagem) : base(id, nome, preco, descricao, comentarios, imagem)
         {
         }
     }
